feat: extract MetricCard to component
